+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.0] - 2025-09-30
+
+### Added
+
+- Added shell completion for bash, zsh, fish, and powershell. Terminal session will have to be started to pickup completion script for the specific shell.
+
+### Changed
+
+- Changed error return code from `99` to `1`.
+
 ## [v0.1.0] - 2025-09-16
 
 Initial release